% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibratedva}
\alias{calibratedva}
\title{Performs Gibbs sampling for calibration}
\usage{
calibratedva(
  A_U,
  A_L = NULL,
  G_L = NULL,
  causes,
  method = c("mshrink", "pshrink"),
  ndraws = 10000,
  burnin = 1000,
  thin = 1,
  ...
)
}
\arguments{
\item{A_U}{When using cause of death predictions from a single algorithm, this
will be a matrix, where each row gives the predicted cause of death probabilities for
an individual death, for individuals without cause of death labels. Each column represents
a cause. If using the top cause, one entry in each row should be 1, while the rest should be 0.
When using predictions from multiple algorithms for the ensemble approach, this should
be a list of matrices with algorithm predictions for the same individuals, where each
entry in the list are predictions from a given algorithm. See examples for more information}

\item{A_L}{A matrix or array in the same format as A_U, but for individuals with labeled
causes of death. If there are no individuals with labeled causes, leave as NULL}

\item{G_L}{A matrix where each row represents either the true cause for an individual
with a labeled cause of death (i.e. if the label for individual i is cause j, then
G_L[i,j] will be 1, and the other entries of that row will be 0), or the probabilities
that each individual died of a certain cause}

\item{causes}{A character vector with the}

\item{method}{One of either "mshrink" (default) for M-shrinkage or "pshrink" for p-shrinkage}

\item{ndraws}{Number of draws in each chain. Default is 10,000}

\item{burnin}{Number of burnin samples. Default is 1,000}

\item{thin}{Thinning parameter. Default is no thinning}

\item{...}{Additional arguments (such as hyperparameter values) passed into
the sampling functions (TO DOCUMENT)}
}
\value{
A list
}
\description{
Takes in estimated causes (or cause probabilities) for both a
representative set of deaths without labels, and an unrepresentative set
of deaths with labels, and estimates the calibrated CSMF
}
