% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{calibva.map}
\alias{calibva.map}
\title{Obtains the MAP estimates for parameters using SQUAREM}
\usage{
calibva.map(test.cod, calib.cod = NULL, calib.truth = NULL, causes,
  epsilon = 0.001, alpha = 5, beta = 0.5, delta = 1,
  init.seed = NULL, tol = 1e-08, maxiter = 50000)
}
\arguments{
\item{test.cod}{will be a vector of length N, with each entry as the estimated
COD (as a character)for indiv. i}

\item{calib.cod}{is in the same format as test.cod, except for the calibration set}

\item{calib.truth}{is a character vector with the true COD for each subject in the
calibration set}

\item{causes}{is a character vector with the names of the causes you are interested in.
The order of the output vector p will correspond to this vector}

\item{epsilon}{A numeric value for the epsilon in the prior}

\item{alpha}{A numeric value for the alpha in the prior}

\item{beta}{A numeric value for the beta in the prior}

\item{delta}{A numeric value for the delta in the prior}

\item{init.seed}{An optional numeric scalar, with the initial seed}
}
\value{
a list, with each object giving the MAP estimates for that parameter
}
\description{
\code{calibva.map} takes in estimation of the underlying
cause of death distribution from training data, as well as a transition
matrix based on calibration data. It uses SQUAREM to obtain MAP estimates of the 
parameters p, M, and gamma
}
