% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibva.sampler}
\alias{calibva.sampler}
\title{Performs Gibbs sampling for calibration}
\usage{
calibva.sampler(test.cod, calib.cod, calib.truth, causes, epsilon, alpha,
  beta, tau.vec, delta, gamma.init, ndraws, max.gamma = 75)
}
\arguments{
\item{test.cod}{will be a vector of length N, with each entry as the estimated
COD (as a character)for indiv. i}

\item{calib.cod}{is in the same format as test.cod, except for the calibration set}

\item{calib.truth}{is a character vector with the true COD for each subject in the
calibration set}

\item{causes}{is a character vector with the names of the causes you are interested in.
The order of the output vector p will correspond to this vector}

\item{epsilon}{A numeric value for the epsilon in the prior}

\item{alpha}{A numeric value for the alpha in the prior}

\item{beta}{A numeric value for the beta in the prior}

\item{tau.vec}{A numeric vector for the logsd for the sampling distributions
of the gammas}

\item{delta}{A numeric value for the delta in the prior}

\item{gamma.init}{A numeric value for the starting value of gammas}

\item{ndraws}{The number of posterior samples to take}

\item{max.gamma}{The maximum value gamma is allowed to take in
posterior samples. Default is 75.}
}
\value{
A list of length \code{ndraws} where each entry in the list contains
the posterior draw for each parameter
}
\description{
\code{calibva.sampler} takes in estimation of the underlying
cause of death distribution from training data, as well as a transition
matrix based on calibration data. Along with the prior values,
it will return a list of posterior samples for parameters of interest
}
