% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calibvaCSMFPosteriorSamples}
\alias{calibvaCSMFPosteriorSamples}
\title{organizes the posterior samples for the CSMF parameters from
\code{calibva.sampler} into a tibble}
\usage{
calibvaCSMFPosteriorSamples(calibva.samples, causes)
}
\arguments{
\item{calibva.samples}{a list returned from \code{calibva.sampler}}

\item{causes}{the cause vector input to CalibVA}
}
\value{
a tibble with each row representing a draw from the posterior sample
of the CSMF for a given cause
}
\description{
organizes the posterior samples for the CSMF parameters from
\code{calibva.sampler} into a tibble
}
