% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calibvaCSMFPosteriorSummary}
\alias{calibvaCSMFPosteriorSummary}
\title{collect CSMF posterior summaries from the CalibVA sampler}
\usage{
calibvaCSMFPosteriorSummary(calibva.samples, causes,
  percentile.L = 0.025, percentile.U = 0.975)
}
\arguments{
\item{calibva.samples}{a list returned from \code{calibva.sampler}}

\item{causes}{the cause vector input to CalibVA}

\item{percentile.L}{the lower percentile for a credible interval. Default .025}

\item{percentile.U}{the upper percentile for a credible interval. Default .975}
}
\value{
a tibble with the posterior means, and confidence intervals of the CSMF and the names of each cause
for each of the draws that are obtained
}
\description{
collect CSMF posterior summaries from the CalibVA sampler
}
