% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibvaIndPredictions}
\alias{calibvaIndPredictions}
\title{Obtains individual COD predictions based on posterior samples}
\usage{
calibvaIndPredictions(calibva.samples, test.cod, causes, burnin = 1000,
  thin = 5)
}
\arguments{
\item{calibva.samples}{a list of posterior samples obtained from \code{calibva.sampler}}

\item{test.cod}{a character vector containing the predicted COD from the algorithm
for each subject in the test set. This should be the same as the argument supplied to \code{calibva.sampler}}

\item{causes}{is a character vector with the names of the causes you are interested in.
This should be the same as the argument supplied to \code{calibva.sampler}}

\item{burnin}{an integer specifying the number of draws you wish to discard
before collecting draws. Default is 1,000}

\item{thin}{an integer specifying the amount the draws should be thinned by}
}
\value{
a list with the first element \code{topCOD} giving the most likely
COD for each individual, based on posterior probability, and the second element
\code{ind.probabilities} giving a matrix such that entry i,j is the posterior
probability of individual i dying from cause j
}
\description{
\code{calibvaIndPredictions} uses posterior samples from the CalibVA
hierarchical bayesian model to obtain the posterior probability for each individual 
dying of given causes, conditional on predictions from a single algorithm
}
