% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gamma_acceptance_rates}
\alias{gamma_acceptance_rates}
\title{computes the acceptance rates for each gamma parameter from a
\code{calibva.sampler} object for each chain}
\usage{
gamma_acceptance_rates(calibva.samples)
}
\arguments{
\item{calibva.samples}{a list returned from \code{calibva.sampler}}
}
\value{
a tibble giving the acceptance rate for each gamma parameter in a chain
}
\description{
computes the acceptance rates for each gamma parameter from a
\code{calibva.sampler} object for each chain
}
