% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRawCSMF}
\alias{getRawCSMF}
\title{Obtain raw CSMF estimates from verbal autopsy guesses}
\usage{
getRawCSMF(test.cod, causes)
}
\arguments{
\item{test.cod}{will be a vector of length N, with each entry as the estimated
COD (as a character)for indiv. i}

\item{causes}{is a character vector with the names of the causes you are interested i}
}
\value{
a data frame with the causes as one column and the CSMF for each cause as 
the second column
}
\description{
Obtain raw CSMF estimates from verbal autopsy guesses
}
