% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs_cod_tanzania}
\alias{gs_cod_tanzania}
\title{PHMRC gold-standard COD for adult deaths in Tanzania}
\format{
A vector giving the PHMRC gold-standard COD for adult deaths in Tanzania.
The order corresonds to the rows in \code{link{tariff_tanzania}}  and \code{link{insilico_tanzania}}
}
\usage{
gs_cod_tanzania
}
\description{
PHMRC gold-standard COD for adult deaths in Tanzania
}
\keyword{datasets}
