% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insilico_tanzania}
\alias{insilico_tanzania}
\title{InSilicoVA probability of COD predictions for adult deaths in Tanzania}
\format{
A matrix with 223 rows and 34 columns. Entry i,j is the predicted probability that
individual i died from cause j. The rownames are the IDs of the deceased.
}
\usage{
insilico_tanzania
}
\description{
A matrix containing predicted probabilities for each COD for adult deaths in Tanzania,
from the PHMRC dataset, using InSilicoVA
}
\keyword{datasets}
