% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_causes}
\alias{map_causes}
\title{Maps cause predictions to broader causes}
\usage{
map_causes(causes, cause_map)
}
\arguments{
\item{causes}{Either a character vector with the single predicted cause per individual,
or a matrix with each row giving cause predictions for an individual. The columns
of this matrix must have the original cause names}

\item{cause_map}{A data frame, with a column \code{cause} giving the original cause,
and a column \code{broad_cause} giving the broader cause corresponding to the 
original cause}
}
\value{
A matrix giving individual probabilities for the broad causes
}
\description{
Maps cause predictions to broader causes
}
