% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_old.R
\name{mle.calibration}
\alias{mle.calibration}
\title{Performs a MLE estimation for the likelihood portion of the CalibVA model}
\usage{
mle.calibration(test.cod, calib.cod, calib.truth, causes)
}
\arguments{
\item{test.cod}{will be a vector of length N, with each entry as the estimated
COD (as a character)for indiv. i}

\item{calib.cod}{is in the same format as test.cod, except for the calibration set}

\item{calib.truth}{is a character vector with the true COD for each subject in the
calibration set}

\item{causes}{is a character vector with the names of the causes you are interested in.
The order of the output vector p will correspond to this vector}
}
\value{
a list with the first element \code{p} being MLE estimates
of the CSMF vector, and the second element \code{M} being the MLE
estimates of the M matrix described in the method
}
\description{
Uses the \link[Rsolnp]{solnp} function to obtain MLE estimates
of p and M, based on the likelihood portion of the CalibVA model
}
