% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mMatrixPosteriorSummary}
\alias{mMatrixPosteriorSummary}
\title{Get the posterior mean and variance for the M matrix}
\usage{
mMatrixPosteriorSummary(calibva.samples, causes,
  output.format = c("matrix", "tibble")[1])
}
\arguments{
\item{calibva.samples}{a mcmc.list object with samples from the CalibVA sampler}

\item{causes}{the cause vector input to CalibVA}

\item{output.format}{either "matrix" or "tibble"}
}
\value{
Either a tibble with the entry wise posterior means or a list of matrices
with the posterior means stored as m_posterior_mean and the posterior variance
stored as m_posterior_variance
}
\description{
Get the posterior mean and variance for the M matrix
}
