% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalizedMisclassificationMatrix}
\alias{normalizedMisclassificationMatrix}
\title{normalize the raw misclassification matrix by row}
\usage{
normalizedMisclassificationMatrix(T.mat)
}
\arguments{
\item{T.mat}{an integer matrix produced by \code{rawMisclassificationMatrix}}
}
\value{
A numeric matrix, where the rows are the conditional misclassification
probabilities
}
\description{
normalize the raw misclassification matrix by row
}
