% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phmrc_adult_cause_map}
\alias{phmrc_adult_cause_map}
\title{Map from PHMRC COD to broad COD}
\format{
A dataframe with 34 rows and 2 columns:
\describe{
  \item{cause}{The PHMRC COD}
  \item{braod_cause}{The broad cause that corresponds to the PHMRC COD}
}
}
\usage{
phmrc_adult_cause_map
}
\description{
Provides a mapping from the 34 causes of death in the PHMRC adult data set (gs_text34)
to 5 broad causes of death
}
\keyword{datasets}
