% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rawMisclassificationMatrix}
\alias{rawMisclassificationMatrix}
\title{obtain the raw misclassification matrix using gold standard and VA COD}
\usage{
rawMisclassificationMatrix(calib.cod, calib.truth, causes)
}
\arguments{
\item{calib.cod}{will be a vector of length N, with each entry as the estimated
COD (as a character) for indiv. i in the calibration set}

\item{calib.truth}{is a character vector with the true COD for each subject in the
calibration set}
}
\value{
an integer matrix T where (i,j)th entry is the number of subjects
in the calibration set with gold standard COD i and VA COD j
}
\description{
obtain the raw misclassification matrix using gold standard and VA COD
}
