% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{revamp.sampler}
\alias{revamp.sampler}
\title{Performs Gibbs sampling for calibration}
\usage{
revamp.sampler(v, T.mat, epsilon, alpha, beta, tau.vec, delta, gamma.init,
  ndraws, max.gamma = 75)
}
\arguments{
\item{v}{a C x 1 matrix, where C is the number of causes. Each row
should contain the number of estimated causes of death from
the training data}

\item{T.mat}{a C x C matrix where the i,j entry is the number of records based
on the calibration set where the true COD is i and the predicted COD is j}

\item{epsilon}{A numeric value for the epsilon in the prior}

\item{alpha}{A numeric value for the alpha in the prior}

\item{beta}{A numeric value for the beta in the prior}

\item{tau.vec}{A numeric vector for the logsd for the sampling distributions
of the gammas}

\item{delta}{A numeric value for the delta in the prior}

\item{gamma.init}{A numeric value for the starting value of gammas}

\item{ndraws}{The number of posterior samples to take}

\item{max.gamma}{The maximum value gamma is allowed to take in
posterior samples. Default is 75.}
}
\value{
A list of length \code{ndraws} where each entry in the list contains
the posterior draw for each parameter
}
\description{
\code{revamp.sampler} takes in estimation of the underlying
cause of death distribution from training data, as well as a transition
matrix based on calibration data. Along with the prior values,
it will return a list of posterior samples for parameters of interest
}
