% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{revamp_with_va}
\alias{revamp_with_va}
\title{Wrapper function for implementing and the ReVAMP calibration}
\usage{
revamp_with_va(train.data, calibration.data, test.data, train.method,
  epsilon, alpha, beta, tau.vec, delta, gamma.init, ndraws,
  max.gamma = 75, nchains = 1, train.seed = NULL,
  sampler.seeds = NULL, ...)
}
\arguments{
\item{train.data}{A data frame/matrix which is compatible for training the given VA method. 
Should have COD labels and be of moderate size}

\item{calibration.data}{A data frame/matrix in the same format as \code{train.data}.}

\item{test.data}{A data frame/matrix in the same format as \code{train.data}
and \code{calibration.data}, except without COD labels.}

\item{train.method}{A character string specifying the VA training method. As of now,
we support "Tariff", "InSilicoVA", and "InterVA".}

\item{epsilon}{A numeric value for the epsilon in the prior}

\item{alpha}{A numeric value for the alpha in the prior}

\item{beta}{A numeric value for the beta in the prior}

\item{tau.vec}{A numeric vector for the logsd for the sampling distributions
of the gammas}

\item{delta}{A numeric value for the delta in the prior}

\item{gamma.init}{A numeric value for the starting value of gammas}

\item{ndraws}{The number of posterior samples to take}

\item{max.gamma}{The maximum value gamma is allowed to take in
posterior samples. Default is 75.}

\item{nchains}{The number of chains for which the ReVAMP sampling method will be run}

\item{sampler.seeds}{An optional vector of numeric values containing the seeds that should be 
set for each chain of the Gibbs sampler. If given, should be of same length as
the value of \code{nchains}}

\item{...}{Additional parameters for the given training method}

\item{train.seeds}{An optional numeric value containing the seed which should be set
before implementing the given training method}
}
\value{
\item{causes }{ The order of the causes in the output of the Gibbs sampler}
\item{v }{ The C x 1 matrix used by the ReVAMP sampler}
\item{T.mat }{ The C x C calibration matrix used by the ReVAMP sampler}
\item{posterior.results}{ A list of length \code{nchains}, which each element
of the list itself being a list containing the output for the ReVAMP sampler
for that chain} \item{train.seed }{ The seed set before implementing the given training method}
\item{sampler.seeds }{ The seeds set before running each chain of the ReVAMP sampler}
}
\description{
\code{revamp_with_va} trains a VA method (currently one of
Tariff, InterVA, or InSilicoVA) on training data (defaults to PHMRC data),
and then uses the calibration and test data to implement the ReVAMP 
hierarchical Bayesian model. This will perform the steps shown in the 
package vignette in one function call
}
\seealso{
\code{\link{revamp.sampler}}
}
