% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{revampCSMF}
\alias{revampCSMF}
\title{collect CSMF estimates from the ReVAMP sampler}
\usage{
revampCSMF(revamp.samples, burnin = 1000, thin = 5)
}
\arguments{
\item{revamp.samples}{a list returned from \code{revamp.sampler}}

\item{burnin}{an integer specifying the number of draws you wish to discard
before collecting draws. Default is 1,000}

\item{thin}{an integer specifying the amount the draws should be thinned by}
}
\value{
a data frame with the values of the CSMF and the names of each cause
for each of the draws that are obtained
}
\description{
\code{revampCSMF} collects draws of the CSMF estimates
after running the \code{revamp.sampler} function, using the
user supplied burn-in and thinning parameters
}
