% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{revampEnsembleIndPredictions}
\alias{revampEnsembleIndPredictions}
\title{Obtains individual COD predictions based on posterior samples from the ensemble ReVAMP method}
\usage{
revampEnsembleIndPredictions(revamp.samples, test.cod.mat, causes,
  burnin = 1000, thin = 5)
}
\arguments{
\item{revamp.samples}{a list of posterior samples obtained from \code{revamp.ensemble.sampler}}

\item{test.cod.mat}{will be a N x K matrix, with entry i,j denoting estimated
COD (as a character)for indiv. i by alg. j This should be the same as the argument 
supplied to \code{revamp.ensemble.sampler}}

\item{causes}{is a character vector with the names of the causes you are interested in.
This should be the same as the argument supplied to \code{revamp.ensemble.sampler}}

\item{burnin}{an integer specifying the number of draws you wish to discard
before collecting draws. Default is 1,000}

\item{thin}{an integer specifying the amount the draws should be thinned by}
}
\value{
a list with the first element \code{topCOD} giving the most likely
COD for each individual, based on posterior probability, and the second element
\code{ind.probabilities} giving a matrix such that entry i,j is the posterior
probability of individual i dying from cause j
}
\description{
\code{revampIndPredictions} uses posterior samples from the ReVAMP
hierarchical bayesian model to obtain the posterior probability for each individual 
dying of given causes, conditional on predictions from multiple algorithms
}
