% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tariff_tanzania}
\alias{tariff_tanzania}
\title{Tariff top COD predictions for adult deaths in Tanzania}
\format{
A data frame with 2023 rows and 2 columns:
\describe{
  \item{ID}{The ID of the deceased}
  \item{cause}{The predicted COD. One of 34 possible causes from gs_text34}
}
}
\usage{
tariff_tanzania
}
\description{
A dataset containing top COD predictions for adult deaths in Tanzania,
from the PHMRC dataset, using Tariff.
}
\keyword{datasets}
