% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{tune_calibratedva}
\alias{tune_calibratedva}
\title{Chooses parameter values for calibration using WAIC}
\usage{
tune_calibratedva(
  A_U,
  A_L = NULL,
  G_L = NULL,
  causes,
  method = c("mshrink", "pshrink"),
  alpha_vec = NULL,
  lambda_vec = NULL,
  which.multimodal = "all",
  which.rhat = "all",
  ...
)
}
\arguments{
\item{A_U}{A matrix or list of causes for individuals without labeled
causes of death. See \code{\link{calibratedva}}.}

\item{A_L}{A matrix or list of causes for individuals with labeled
causes of death. See \code{\link{calibratedva}}.}

\item{G_L}{A matrix where each row represents either the true cause for an individual
with a labeled cause of death. See \code{\link{calibratedva}}.}

\item{causes}{A character vector with the names of the causes. These should correspond to
the columns of \code{A_U}, \code{A_L}, and \code{G_L}}

\item{method}{One of either "mshrink" (default) for M-shrinkage or "pshrink" for p-shrinkage}

\item{alpha_vec}{If using M-shrinkage vector of values for alpha for which the function will
evaluate the WAIC. If not provided, the function will auto-generate a vector.}

\item{lambda_vec}{If using p-shrinkage vector of values for lambda for which the function will
evaluate the WAIC. If not provided, the function will auto-generate a vector.}

\item{which.multimodal}{A character specifying whether both p and M (which.multimodal = "all")
should be evaluated for multimodality, or just p (which.multimodal = "p")}

\item{which.rhat}{A character specifying whether both p and M (which.rhat = "all")
should be evaluated for convergence, or just p (which.rhat = "p")}

\item{...}{Additional arguments passed into \code{\link{calibratedva}}.}
}
\value{
A list with the following components.
\describe{
  \item{final_model}{A list with output specified in
  \code{\link{calibratedva}} object containing the posterior samples
  for the best value of the parameter (either alpha or lambda)}
  \item{alpha_final}{The chosen value of alpha for the posterior samples.
  Null if using p-shrinkage.}
  \item{lambda_final}{The chosen value of lambda for the posterior samples.
  Null if using M-shrinkage.}
  \item{waic_df}{A data frame containing a row for each parameter evaluated,
  which gives the WAIC, whether or not the posteriors were multimodal, and the
  Rhat of the posterior samples.}
}
}
\description{
Uses a grid search to choose the shrinkage parameter for calibration
which gives the lowest WAIC values.
}
