% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibratedva}
\alias{calibratedva}
\title{Performs Gibbs sampling for calibration}
\usage{
calibratedva(
  va_unlabeled,
  va_labeled = NULL,
  gold_standard = NULL,
  causes,
  method = c("mshrink", "pshrink"),
  nchains = 3,
  ndraws = 10000,
  burnin = 1000,
  thin = 1,
  pseudo_samplesize = 100,
  alpha = 5,
  beta = 0.5,
  lambda = 1,
  delta = 1,
  epsilon = 0.001,
  tau = 0.5,
  which.multimodal = "all",
  which.rhat = "all",
  print.chains = FALSE,
  init.seed = 123
)
}
\arguments{
\item{va_unlabeled}{When using cause of death predictions from a single algorithm, this
will be a matrix, where each row gives the predicted cause of death probabilities for
an individual death, for individuals without cause of death labels. Each column represents
a cause. If using the top cause, one entry in each row should be 1, while the rest should be 0.
When using predictions from multiple algorithms for the ensemble approach, this should
be a list of matrices with algorithm predictions for the same individuals, where each
entry in the list are predictions from a given algorithm. See examples for more information}

\item{va_labeled}{A matrix or list in the same format as va_unlabeled, but for individuals with labeled
causes of death. If there are no individuals with labeled causes, leave as NULL}

\item{gold_standard}{A matrix where each row represents either the true cause for an individual
with a labeled cause of death (i.e. if the label for individual i is cause j, then
gold_standard[i,j] will be 1, and the other entries of that row will be 0), or the probabilities
that each individual died of a certain cause. The rows of \code{G_L} should correspond
to the rows of \code{A_L} (or the rows of each element of \code{A_L} if it is a list)}

\item{causes}{A character vector with the names of the causes. These should correspond to
the columns of \code{A_U}, \code{A_L}, and \code{G_L}}

\item{method}{One of either "mshrink" (default) for M-shrinkage or "pshrink" for p-shrinkage}

\item{nchains}{The number of chains. Default is 3}

\item{ndraws}{Number of draws in each chain. Default is 10,000}

\item{burnin}{Number of burnin samples. Default is 1,000}

\item{thin}{Thinning parameter. Default is no thinning}

\item{pseudo_samplesize}{The number of pseudo samples (T) used for the 
Gibbs Sampler using rounding and coarsening. Default is 100.}

\item{alpha}{A numeric value for the alpha in the prior of gamma when using M-shrinkage.
Higher values (relative to beta) leads to more shrinkage. Default is 5. If using
the ensemble model, a vector of length K can be used (where K is the number of algorithms).}

\item{beta}{A numeric value for the beta in the prior of gamma when using M-shrinkage.
Default is .5.}

\item{lambda}{A numeric value for the lambda in the prior of p for p-shrinkage.
Higher values leads to more shrinkage. Default is 1.
#' @param delta A numeric value for the delta in the prior of p. Only used for
M-shrinkage sampling.}

\item{epsilon}{A numeric value for the epsilon in the prior of M. Default is .001.}

\item{which.multimodal}{A character specifying whether both p and M (which.multimodal = "all")
should be evaluated for multimodality, or just p (which.multimodal = "p")}

\item{which.rhat}{A character specifying whether both p and M (which.rhat = "all")
should be evaluated for convergence, or just p (which.rhat = "p")}

\item{print.chains}{A logical scalar which says whether or not you want the 
progress of the sampling printed to the screen. Default is FALSE}

\item{init.seed}{The initial seed for sampling. Default is 123.}

\item{tau.vec}{A numeric vector for the log standard deviation for the sampling distributions
of the gammas. Only used for M-shrinkage sampling.}
}
\value{
A list with the following components.
\describe{
  \item{samples}{A \code{\link[coda]{mcmc.list}} object containing the posterior samples
  for p, M, and gamma (if using M-shrinkage)}
  \item{A_U}{The value of \code{va_unlabeled} using for the posterior samples}
  \item{A_L}{The value of \code{va_labeled} using for the posterior samples}
  \item{G_L}{The value of \code{gold_standard} using for the posterior samples}
  \item{method}{The method used for shrinkage (either mshrink or pshrink)}
  \item{waic}{The estimated WAIC for the calibrated posterior}
  \item{waic_uncalib}{The estimated WAIC for the uncalibrated posterior}
  \item{multimodal}{Either TRUE or FALSE, indicating whether or not the posterior
  samples for p (and potentially M) are multimodal}
  \item{rhat_max}{The maximum rhat for p (and potentially M), which can be used
  for evaluating convergence}
  \item{alpha}{The value(s) of alpha used (if method = "mshrink")}
  \item{beta}{The value of beta used (if method = "mshrink")}
  \item{lambda}{The value of lambda used (if method = "pshrink")}
}
}
\description{
Takes in estimated causes (or cause probabilities) for both a
representative set of deaths without labels, and an unrepresentative set
of deaths with labels, and estimates the calibrated CSMF
}
